pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "red-forge"
version = "0.1.0"
description = "AI-Powered Regulatory Compliance Monitoring Agent"
requires-python = ">=3.11"

dependencies = [
    # LangChain + LangGraph
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langsmith>=0.1.0",

    # Google Vertex AI
    "google-cloud-aiplatform>=1.60.0",
    "vertexai>=1.60.0",

    # Vector Store
    "qdrant-client>=1.10.0",

    # Knowledge Graph
    "neo4j>=5.20.0",

    # Database
    "psycopg[binary]>=3.2.1",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",

    # Document Parsing
    "docling>=2.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # Pipeline Orchestration
    "prefect>=3.0.0",
    "prefect-sqlalchemy>=0.4.0",

    # Storage
    "minio>=7.2.0",
    "boto3>=1.34.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Data & Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Monitoring
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",

    # Utilities
    "tenacity>=8.3.0",
    "python-dateutil>=2.9.0",
    "feedparser>=6.0.11",
    "requests>=2.32.0",
    "rich>=13.7.0",
    "structlog>=24.2.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "knowledge", "agents", "storage", "org_context", "api"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
